Simplify Boost installation
Everything is installed in same file which is run for every Python version to simplify installation process and remove duplicated code. However, there are checks for already installed libraries so there won't be any unnecessary duplicated building and installing.

+++ bindings/python/tools/install_boost.sh
@@ -1,35 +1,56 @@
-#! /bin/bash
+#!/bin/bash
 
 set -e
 set -x
 
-# Don't re-build and re-install Boost if it was already installed previously
-if [[ "$OSTYPE" == "msys" ]] && [[ -f /c/Boost/bin/b2.exe ]]
+# Get number of CPU cores and Python version
+cores=$(python -c "import multiprocessing; print(multiprocessing.cpu_count(), end='')")
+version=$(python -c "import sys; print(sys.version[:3].replace('.', ''), end='')")
+model=$(python -c "import sys; print('x64' if sys.maxsize > 2**32 else 'x32', end='')")
+
+# Write Python config to fix bug where Boost guesses wrong paths
+python $1/bindings/python/tools/generate_boost_config.py ~/user-config.jam
+
+# Don't re-build and re-install Boost if it was already installed
+# If both Boost.Boost and Boost.Python are already installed, there isn't any need for additional Boost.System check
+if [[ "$OSTYPE" == "msys" ]] && [[ -f /c/Boost/bin/b2.exe ]] && [[ -f /c/Boost/lib/libboost_python$version-mt-$model.lib ]]
 then
   echo Using cache...
+  rm -f /c/Boost/lib/boost_system.lib /c/Boost/lib/boost_python3.lib
+  ln /c/Boost/lib/libboost_system-mt-$model.lib /c/Boost/lib/boost_system.lib
+  ln /c/Boost/lib/libboost_python$version-mt-$model.lib /c/Boost/lib/boost_python3.lib
   exit 0
-elif [[ "$OSTYPE" == "darwin"* ]] && [[ -f /usr/local/bin/b2 ]]
+elif [[ "$OSTYPE" == "darwin"* ]] && [[ -f /usr/local/bin/b2 ]] && [[ -f /usr/local/lib/libboost_python$version-mt.dylib ]]
 then
   echo Using cache...
   exit 0
 fi
 
-# Download Boost sources
-curl -L https://dl.bintray.com/boostorg/release/1.73.0/source/boost_1_73_0.tar.gz -o /tmp/boost.tar.gz
-tar xzf /tmp/boost.tar.gz -C /tmp
-
-# Install Boost from Homebrew on macOS
-if [[ "$OSTYPE" == "darwin"* ]]
+# Install Boost from Homebrew on macOS if it was not already installed
+if [[ "$OSTYPE" == "darwin"* ]] && [[ ! -f /usr/local/bin/b2 ]]
 then
   brew update
   brew install boost boost-build
-  exit 0
 fi
 
-# Add modern Python to PATH on Linux
-if [[ "$OSTYPE" == "linux-gnu"* ]]
+# Download Boost sources if they are not already downloaded
+# They are needed for Boost.Python libraries, even on macOS
+if [[ ! -d /tmp/boost_1_73_0 ]]
 then
-  PATH=/opt/python/cp38-cp38/bin:$PATH
+  curl -L https://dl.bintray.com/boostorg/release/1.73.0/source/boost_1_73_0.tar.gz -o /tmp/boost.tar.gz
+  tar xzf /tmp/boost.tar.gz -C /tmp
+fi
+
+# Install Boost.Python as root and use same config as Boost from Homebrew on macOS
+# On other systems, just use normal tagged layout for whole Boost installed
+if [[ "$OSTYPE" == "darwin"* ]]
+then
+  root=sudo
+  threading=threading=multi,single
+  link=link=shared,static
+  layout=--layout=tagged-1.66
+else
+  layout=--layout=tagged
 fi
 
 # Use MSVC and specific prefixes for Boost on Windows
@@ -39,16 +60,47 @@ then
   prefix=--prefix=C:/Boost
 fi
 
-# Get number of CPU cores
-cores=$(python -c "import multiprocessing; print(multiprocessing.cpu_count(), end='')")
+# Don't run on macOS because it is already installed from Homebrew
+# Also don't run if it was already installed on previous run
+if [[ "$OSTYPE" != "darwin"* ]] && ! ([[ -f /usr/local/bin/b2 ]] || [[ -f /c/Boost/bin/b2.exe ]])
+then
+  # Install Boost.System
+  cd /tmp/boost_1_73_0
+  rm -f project-config.jam
+  ./bootstrap.sh --with-libraries=system $prefix
+  ./b2 install release $toolset $threading $link $layout $prefix -j$cores
 
-# Install Boost System
+  # Install Boost.Build
+  cd /tmp/boost_1_73_0/tools/build
+  ./bootstrap.sh
+  ./b2 install release $toolset -j$cores
+fi
+
+# Install Boost.Python
+# Check if this version was already installed is not needed, because it happens at the start of the file
 cd /tmp/boost_1_73_0
 rm -f project-config.jam
-./bootstrap.sh --with-libraries=system $prefix
-./b2 install release $toolset --layout=tagged $prefix -j$cores
+./bootstrap.sh --with-libraries=python --with-python=python3 $prefix
+$root ./b2 install release $toolset $threading $link $layout $prefix -j$cores
 
-# Install Boost Build
-cd /tmp/boost_1_73_0/tools/build
-./bootstrap.sh
-./b2 install release $toolset -j$cores
+# Link Boost.System and Boost.Python libraries to correct name
+if [[ "$OSTYPE" == "linux-gnu"* ]]
+then
+  rm -f /usr/local/lib/libboost_system.a /usr/local/lib/libboost_system.so
+  rm -f /usr/local/lib/libboost_python3.a /usr/local/lib/libboost_python3.so
+  ln /usr/local/lib/libboost_system-mt-$model.a /usr/local/lib/libboost_system.a
+  ln /usr/local/lib/libboost_system-mt-$model.so /usr/local/lib/libboost_system.so
+  ln /usr/local/lib/libboost_python$version-mt-$model.a /usr/local/lib/libboost_python3.a
+  ln /usr/local/lib/libboost_python$version-mt-$model.so /usr/local/lib/libboost_python3.so
+elif [[ "$OSTYPE" == "msys" ]]
+then
+  rm -f /c/Boost/lib/boost_system.lib /c/Boost/lib/boost_python3.lib
+  ln /c/Boost/lib/libboost_system-mt-$model.lib /c/Boost/lib/boost_system.lib
+  ln /c/Boost/lib/libboost_python$version-mt-$model.lib /c/Boost/lib/boost_python3.lib
+fi
+
+# Fix Boost.Python library ID on macOS
+if [[ "$OSTYPE" == "darwin"* ]]
+then
+  $root install_name_tool /usr/local/lib/libboost_python$version-mt.dylib -id /usr/local/lib/libboost_python$version-mt.dylib
+fi
